First Three Sections
The first three sections are responding appropriately in both Safari
and Chrome; the behavior isn’t identical but I think it’s acceptable
(notably the outline of Scotland).

I’d like to figure out a way to switch the text-justification from
right to left when the column breaks to the next line. The
justify-right looks good when the columns are side-by-side, but as soon
as the text flows under it looks strange. I haven’t written a query for
that yet.